backend/requirements.txt
# FastAPI Backend Dependencies for NavEdge Phase 2
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
supabase==2.0.2

# AI & Image Processing
opencv-python==4.8.1.78
pillow==10.1.0
pytesseract==0.3.10
scikit-image==0.22.0
numpy==1.24.3

# Advanced AI Models
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
lpips>=0.1.4

# AWS S3 Storage
boto3>=1.28.0
botocore>=1.31.0

# Additional ML dependencies
matplotlib>=3.7.0
seaborn>=0.12.0

# PDF Generation
reportlab==4.0.7
weasyprint==60.2

# HTTP Requests & Web Scraping
httpx==0.25.2
beautifulsoup4==4.12.2
selenium==4.15.2

# WhatsApp Integration
twilio==8.10.0
requests==2.31.0

# Background Jobs
apscheduler==3.10.4
celery==5.3.4
redis==5.0.1

# Email
smtplib
email-validator==2.1.0

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
python-dateutil==2.8.2
pandas==2.1.4
openpyxl==3.1.2

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
